refactor: BatchEnrollButton 컴포넌트 분리
